requirements.txt
python-telegram-bot==20.*
python-dotenv>=1.0.0
flask>=2.0.0
gunicorn>=21.0.0